REFACTOR | Refactor code to make it easier to add new features
