+++ public/index.html
@@ -129,8 +129,8 @@
 <script type="text/javascript" src="packages/socket.io.js"></script>
 <script type="text/javascript" src="packages/jquery.json-viewer-master/json-viewer/jquery.json-viewer.js"></script>
 
-<script type="text/javascript" src="src/main.js"></script>
-<script type="text/javascript" src="src/resend.js"></script>
-<script type="text/javascript" src="src/settings.js"></script>
+<script type="text/javascript" src="client/src/main.js"></script>
+<script type="text/javascript" src="client/src/resend.js"></script>
+<script type="text/javascript" src="client/src/settings.js"></script>
 
 </html>
